+++ CHANGELOG.md
@@ -3,6 +3,7 @@ All notable changes to `knotx-data-bridge` will be documented in this file.
 
 ## Unreleased
 List of changes that are finished but not yet released in any final version.
+- [PR-37](https://github.com/Knotx/knotx-data-bridge/pull/37) - implementation of [fallback handling](https://github.com/Cognifide/knotx/issues/466) in data bridge
 
 ## 1.4.0
 Initial open source release.